lib/requirements.txt
biopython==1.79
biothings-client==0.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cycler==0.11.0
datrie==0.8.2
dbus-python==1.2.18
distlib==0.3.4
filelock==3.4.2
fonttools==4.28.5
idna==3.3
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-venn==0.11.6
mygene==3.2.2
natsort==8.0.2
numpy==1.22.1
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==9.0.0
platformdirs==2.4.1
ply==3.11
pycparser==2.21
pycurl==7.44.1
PyGObject==3.42.0
pyparsing==3.0.6
python-apt==2.3.0+b1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
rpy2==3.4.5
scikit-posthocs==0.6.7
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
statsmodels==0.13.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
virtualenv==20.13.0
xmltramp2==3.1.1
